requirements-dev.txt
# Include all regular dependencies
-r requirements.txt

# UI Testing
pexpect==4.8.0
pytest-console-scripts==1.4.1

# Mock server
fogis-api-client-timmyBird[mock]>=0.5.0
